[https://trac.osgeo.org/mapguide/ticket/1657]: Add/Fix support for 64-bit Linux. With this change, we have a functional MapGuide Server on 64-bit linux. Some FDO providers may still not be functional yet on 64-bit Linux (eg. SHP) and will have to be resolved on a provider-by-provider basis. At least with this change, we can still pave the way for a MapGuide release for 64-bit Linux with partial FDO provider support.
The typedef changes has somehow caused static members in MgOrderingOption and MgFeatureSpatialOperations to be unresolved on Linux. This submission adds .cpp files to define these members ensuring these static members are included in the final library .so

Reviewed by Walt Welton-Lair.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@8399 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/Stylization/RS_Raster.h
@@ -21,6 +21,7 @@
 #include "RS_InputStream.h"
 #include "Bounds.h"
 #include "RendererStyles.h"
+#include <stdint.h>
 
 class RS_Raster
 {
@@ -44,7 +45,7 @@ public:
     virtual RS_String       GetVerticalUnits()  = 0;
 
     virtual int             GetNullValueType()  = 0;
-    virtual long long       GetNullValueData()  = 0;
+    virtual int64_t         GetNullValueData()  = 0;
 };
 
 #endif
